add hadle platform response
