+++ app/util/requests.ts
@@ -15,7 +15,7 @@ export const handleResponse = (response: Response): Promise<any> => {
   const contentType = response.headers.get('content-type');
   const isJson = contentType && contentType.includes('application/json');
   const result = isJson ? response.json() : response.text();
-  return response.ok ? result : Promise.reject(response);
+  return response.ok ? result : result.then(r => Promise.reject(r));
 };
 
 export const handleErrors = (response: Response): Promise<any> => {
@@ -33,25 +33,6 @@ export function camelize(response: Response): Promise<any> {
       resolve(humps.camelizeKeys(json));
     });
   });
-}
-
-export function requiresToken() {
-  return (target: any, methodName: string, descriptor: PropertyDescriptor) => {
-    const original = descriptor.value;
-    return {
-      ...descriptor,
-      value(...args: any[]) {
-        return original.apply(target.constructor.instance, args).catch((error: Response) => {
-          if (error.status === 401) {
-            return target.fetchNewToken().then(() => {
-              return original.apply(target.constructor.instance, args);
-            });
-          }
-          return Promise.reject(error);
-        });
-      },
-    };
-  };
 }
 
 /**
